ARpdb written, used only in netcat (so far)
