+++ Pipfile
@@ -1,0 +1,14 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+flake8 = "*"
+pylint = "*"
+
+[packages]
+pdbpp = "*"
+
+[requires]
+python_version = "3.7"
