+++ rpdb/rpdb.py
@@ -1,0 +1,128 @@
+"""Remote Python Debugger (pdb wrapper)."""
+
+__author__ = "Bertrand Janin <b@janin.com>"
+__version__ = "0.1.6"
+
+import pdb
+import socket
+import threading
+import sys
+
+
+class FileObjectWrapper(object):
+    def __init__(self, fileobject, stdio):
+        self._obj = fileobject
+        self._io = stdio
+
+    def __getattr__(self, attr):
+        if hasattr(self._obj, attr):
+            attr = getattr(self._obj, attr)
+        elif hasattr(self._io, attr):
+            attr = getattr(self._io, attr)
+        else:
+            raise AttributeError("Attribute %s is not found" % attr)
+        return attr
+
+
+class Rpdb(pdb.Pdb):
+
+    def __init__(self, addr, port):
+        """Initialize the socket and initialize pdb."""
+
+        # Backup stdin and stdout before replacing them by the socket handle
+        self.old_stdout = sys.stdout
+        self.old_stdin = sys.stdin
+        self.port = port
+
+        # Open a 'reusable' socket to let the webapp reload on the same port
+        self.skt = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        self.skt.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, True)
+        self.skt.bind((addr, port))
+        self.skt.listen(1)
+
+        # Writes to stdout are forbidden in mod_wsgi environments
+        try:
+            sys.stderr.write("pdb is running on %s:%d\n"
+                             % self.skt.getsockname())
+        except IOError:
+            pass
+
+        (clientsocket, address) = self.skt.accept()
+        handle = clientsocket.makefile('rw')
+        pdb.Pdb.__init__(self, completekey='tab',
+                         stdin=FileObjectWrapper(handle, self.old_stdin),
+                         stdout=FileObjectWrapper(handle, self.old_stdin))
+        sys.stdout = sys.stdin = handle
+        self.handle = handle
+        OCCUPIED.claim(port, sys.stdout)
+
+    def shutdown(self):
+        """Revert stdin and stdout, close the socket."""
+        sys.stdout = self.old_stdout
+        sys.stdin = self.old_stdin
+        self.handle.close()
+        OCCUPIED.unclaim(self.port)
+        self.skt.shutdown(socket.SHUT_RDWR)
+        self.skt.close()
+
+    def do_continue(self, arg):
+        """Clean-up and do underlying continue."""
+        try:
+            return pdb.Pdb.do_continue(self, arg)
+        finally:
+            self.shutdown()
+
+    do_c = do_cont = do_continue
+
+    def do_quit(self, arg):
+        """Clean-up and do underlying quit."""
+        try:
+            return pdb.Pdb.do_quit(self, arg)
+        finally:
+            self.shutdown()
+
+    do_q = do_exit = do_quit
+
+    def do_EOF(self, arg):
+        """Clean-up and do underlying EOF."""
+        try:
+            return pdb.Pdb.do_EOF(self, arg)
+        finally:
+            self.shutdown()
+
+
+class OccupiedPorts(object):
+    """Maintain rpdb port versus stdin/out file handles.
+
+    Provides the means to determine whether or not a collision binding to a
+    particular port is with an already operating rpdb session.
+
+    Determination is according to whether a file handle is equal to what is
+    registered against the specified port.
+    """
+
+    def __init__(self):
+        self.lock = threading.RLock()
+        self.claims = {}
+
+    def claim(self, port, handle):
+        self.lock.acquire(True)
+        self.claims[port] = id(handle)
+        self.lock.release()
+
+    def is_claimed(self, port, handle):
+        self.lock.acquire(True)
+        got = (self.claims.get(port) == id(handle))
+        self.lock.release()
+        return got
+
+    def unclaim(self, port):
+        self.lock.acquire(True)
+        del self.claims[port]
+        self.lock.release()
+
+
+# {port: sys.stdout} pairs to track recursive rpdb invocation on same port.
+# This scheme doesn't interfere with recursive invocations on separate ports -
+# useful, eg, for concurrently debugging separate threads.
+OCCUPIED = OccupiedPorts()
