+++ test.py
@@ -1,0 +1,11 @@
+import rpdb
+
+print("aaaa")
+
+print("bbbb")
+
+rpdb.set_trace(active=True)
+
+c = 20
+
+d = print("dddd")
